Add Authentication (Spring Security, Configuration, Login, Filter, JWT, And Exception Handling) & Authorization For Posts

+++ src/main/java/com/playground/springbootblogapp/post/PostService.java
@@ -1,10 +1,13 @@
 package com.playground.springbootblogapp.post;
 
+import com.playground.springbootblogapp.authentication.AuthenticationService;
 import com.playground.springbootblogapp.user.User;
 import com.playground.springbootblogapp.user.UserRepository;
+import com.playground.springbootblogapp.user.UserService;
 import lombok.AllArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
@@ -15,7 +18,7 @@ import java.util.UUID;
 public class PostService {
     private final PostRepository postRepository;
     private final PostMapper postMapper;
-    private final UserRepository userRepository;
+    private final AuthenticationService authenticationService;
 
     public Page<PostDto> getPageablePosts(Pageable pageable) {
         return postRepository.findAll(pageable).map(postMapper::toDto);
@@ -28,7 +31,7 @@ public class PostService {
     }
 
     public PostDto createPost(CreatePostRequest request) {
-        User user = userRepository.findById(request.userId()).orElseThrow();
+        User user = authenticationService.getCurrentUser();
 
         Post post = postMapper.toEntity(request);
         post.setCreated(LocalDateTime.now());
@@ -41,6 +44,11 @@ public class PostService {
 
     public PostDto editPost(UUID uuid, UpdatePostRequest request) {
         Post post = postRepository.findById(uuid).orElseThrow(PostNotFoundException::new);
+        User user = authenticationService.getCurrentUser();
+
+        if (post.getUser() != user) {
+            throw new AccessDeniedException("You can not edit other people's posts");
+        }
 
         postMapper.update(request, post);
 
@@ -50,7 +58,13 @@ public class PostService {
     }
 
     public void deletePost(UUID uuid) {
-        postRepository.findById(uuid).orElseThrow(PostNotFoundException::new);
+        Post post = postRepository.findById(uuid).orElseThrow(PostNotFoundException::new);
+        User user = authenticationService.getCurrentUser();
+
+        if (post.getUser() != user) {
+            throw new AccessDeniedException("You can not delete other people's posts");
+        }
+
         postRepository.deleteById(uuid);
     }
 }
